REF add more tests and break up big files
