+++ ngmix/priors/__init__.py
@@ -1,3 +1,6 @@
 # flake8: noqa
+from .kde import *
+from .multivariate import *
 from .priors import *
 from .random import make_rng, srandu
+from .shape import *
